+++ CHANGES.rst
@@ -9,6 +9,7 @@ Unreleased
 * Add method :meth:`~.Subreddits.premium` to reflect the naming change in
   Reddit's API.
 * Ability to submit image galleries with :meth:`.submit_gallery`.
+* Ability to pass a gallery url to :meth:`.Reddit.submission`.
 * Ability to specify modmail mute duration.
 
 **Changed**